Remove the rumor heat mechanism.
The rumor heat mechanism described in the Xerox paper that our rumor
implementation is based on uses the heat to determine how many times
a rumor has been sent to the network as a whole. Our implementation
didn't do that. Instead, it used the heat to determine how many times
a rumor had been sent to each member of the network.

It turns out that's not necessary because we're using zmq as
a transport for the rumors, which uses TCP under the hood, so delivering
a rumor once is sufficient.

Now we just send out any rumors we have that aren't expired, which is
closer to an active anti-entropy approach than rumor mongering. Note that
this is still sub-optimal, because we're sending rumors that our
neighbors could potentially already have. In theory, network traffic
will be somewhat reduced, given that there will be less rumors to send,
and we will only be sending rumors once.

There's plenty of room left here for optimizations, but those are out of
scope for this particular task.

Signed-off-by: Josh Black <raskchanky@gmail.com>